Created LinkedList and Node classes with tests passing
